requirements.txt
keyboard>=0.13.5
pyperclip>=1.8.2
requests>=2.31.0
beautifulsoup4>=4.12.0
pystray>=0.19.0
Pillow>=10.0.0
